+++ packages/http/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @openfn/language-http
 
+## 5.0.4
+
+### Patch Changes
+
+- Updated dependencies [c19efbe]
+  - @openfn/language-common@1.11.1
+
+## 5.0.3
+
+### Patch Changes
+
+- Updated dependencies [85c35b8]
+  - @openfn/language-common@1.11.0
+
 ## 5.0.2
 
 ### Patch Changes