+++ packages/openfn/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @openfn/language-openfn
 
+## 1.3.8
+
+### Patch Changes
+
+- Updated dependencies [c19efbe]
+  - @openfn/language-common@1.11.1
+
+## 1.3.7
+
+### Patch Changes
+
+- Updated dependencies [85c35b8]
+  - @openfn/language-common@1.11.0
+
 ## 1.3.6
 
 ### Patch Changes